% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_sightings}
\alias{monthly_sightings}
\title{Monthly number of UFO sightings for the U.S.}
\format{
A data frame with 520 rows and 4 columns.

\describe{
  \item{date}{date of sighting, rounded down to first day of the month}
  \item{sightings}{number of sightings in that month}
  \item{year}{year of the sighting}
  \item{month}{month of the sighting}
}
}
\usage{
monthly_sightings
}
\description{
A data set containing number of monthly UFO sightings for the U.S. since
 1970.
}
\keyword{datasets}
