% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_sightings_by_shape}
\alias{monthly_sightings_by_shape}
\title{Monthly number of UFO sightings for each shape.}
\format{
A data frame with 14588 rows and 5 columns.

\describe{
  \item{date}{date of sighting, rounded down to first day of the month}
  \item{shape}{the shape of the UFO}
  \item{sightings}{number of sightings for that shape and month}
  \item{year}{year of the sighting}
  \item{month}{month of the sighting}
}
}
\usage{
monthly_sightings_by_shape
}
\description{
A data set containing number of monthly UFO sightings for each shape.
}
\keyword{datasets}
