% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape_counts}
\alias{shape_counts}
\title{Number of sightings for each UFO shape.}
\format{
A data frame with 28 rows and 2 columns.

\describe{
  \item{shape}{the shape of the sighting}
  \item{sightings}{number of sightings for that shape}
}
}
\usage{
shape_counts
}
\description{
A data set containing the number of sightings for each UFO shape.
}
\keyword{datasets}
