% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sightings}
\alias{sightings}
\title{UFO sightings in each U.S. state.}
\format{
A data frame with 63496 rows and 2 columns.

\describe{
  \item{date}{date of sighting}
  \item{state}{state where sighting occured}
}
}
\source{
\url{https://www.kaggle.com/NUFORC/ufo-sightings}
}
\usage{
sightings
}
\description{
A data set containing the date and state of UFO sightings since 1970.
}
\keyword{datasets}
