% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_sightings_by_state}
\alias{total_sightings_by_state}
\title{UFO sighting counts by U.S. state}
\format{
A simple features data frame with 51 rows and 3 columns.

\describe{
  \item{state}{state where sighting occured}
  \item{sightings}{number of sightings in that state}
  \item{geometry}{simple features list-column containing state boundaries}
  }
}
\usage{
total_sightings_by_state
}
\description{
A data set containing UFO sighting counts for each state. Also contains
 simple features needed for plotting.
}
\keyword{datasets}
